Made some changes to picture caching along with updates based on tester's feedback

+++ Saloof/Saloof/DPImageCache.swift
@@ -16,11 +16,13 @@ class DPImageCache: NSObject {
             .UserDomainMask,
             true)[0] as! NSString).stringByAppendingPathComponent(CACHEPATH)
         let fileManage: NSFileManager = NSFileManager.defaultManager()
-        var allFiles: Array = fileManage.contentsOfDirectoryAtPath(cachePath, error: nil)!
-        for object in enumerate(allFiles) {
-            fileManage.removeItemAtPath(cachePath.stringByAppendingPathComponent(
-                object.element as! String),
-                error: nil)
+        if fileManage.contentsOfDirectoryAtPath(cachePath, error: nil)?.count != nil{
+            var allFiles: Array = fileManage.contentsOfDirectoryAtPath(cachePath, error: nil)!
+            for object in enumerate(allFiles) {
+                fileManage.removeItemAtPath(cachePath.stringByAppendingPathComponent(
+                    object.element as! String),
+                    error: nil)
+            }
         }
     }
 }
